address review: move previewUrl cleanup to useOpenClawAttachments, add default params to hooks
- Move object URL revocation on unmount from useOpenClawStream into
  useOpenClawAttachments where the URLs are created (fixes hidden coupling)
- Remove attachmentsRef tracking from useOpenClawStream (no longer needed)
- Add `= {}` default to useOpenClawStream and useOpenClawAttachments signatures
- Add `onError = () => {}` and `onSendComplete = () => {}` defaults to
  useOpenClawStream to match repo hook conventions and avoid runtime errors
